webScraper2.py
"""
program will use urllib to read HTML from a data file,
it will extract href= values from the anchor tags, scan 
for a tag that is in a particular position relative to 
the first name in the list, follow that link and repeat
the process and report the last name you find.

Find link at pos = 3 (first name is pos = 1). follow link, 
and repeat 4 times. Answer is last name you retrieve.   

"""